Improve the huge allocator metadata management.
Rather than trying to keep two sets of metadata in sync, iterate over
huge allocators by iterating over all of the objects owned by the huge
allocator allocator and filter out ones that do not own an allocation
(i.e. ones in the early stages of alloc or the late stages of free).