Initial support for free().
free() lets the GC run as a temporal memory management bug checker: A
freed object is not deleted, it is simply marked as an object that
should be unreachable.  If the object is still reachable, then this is
likely a bug.

Currently, we can only report that an object is reachable, not from
where it is reachable (which would be more useful).

We are also not able to differentiate reachable-from-live-stack-values from
reachable-from-leftover-stack-values.  There are unlikely to be a
significant number of valid capabilities left on the stack if we only
walk up, because most stack frames are overwritten (though buffers that
are passed down the stack by pointer are the exception).  We need to run
some experiments to see the extent to which this is true.